pyproject.toml
[project]
name = "aisprint_ams_client"
description = "AI-SPRINT Monitoring Subsystem (AMS) Python client library"
version = "1.0"
keywords = [ "aisprint", "AI-SRINT", "AMS" ]
authors = [
    { name="Michał Soczewka", email="msoczewka@7bulls.com" }
]
maintainers = [
    { name="Michał Soczewka", email="msoczewka@7bulls.com" }
]
readme = "README.txt"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Benchmark",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Logging",
    "Operating System :: OS Independent"
]
requires-python = ">=3.8"
dependencies = [
    "pytelegraf==0.3.3",
    "fluent-logger==0.10.0"
]

[project.urls]
Homepage = "https://www.ai-sprint-project.eu/"
Repository = "https://github.com/ai-sprint-eu-project"

[build-system]
requires = ["setuptools>=56.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["aisprint*"]
